feat(auth/frontend): use /userinfo endpoint and rename RTK query elements
