+++ frontend/src/features/auth/index.ts
@@ -1,2 +1,2 @@
-export { selectSessionResult, selectUser } from "./authSlice"
+export { selectUserInfoResult, selectUserFullName } from "./authSlice"
 export { login, logout } from "./authActions"
