+++ frontend/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export { default as useSession } from "./useSession"
+export { default as useUserInfo } from "./useUserInfo"
