+++ frontend/src/routes/PrivateRoute.tsx
@@ -2,16 +2,16 @@ import React from "react"
 import { Navigate } from "react-router-dom"
 
 import { history } from "./history"
-import { useSession } from "../hooks"
+import { useUserInfo } from "../hooks"
 
 function PrivateRoute({ children }) {
-  const { session, isLoading, isError } = useSession()
+  const { userInfo, isLoading, isError } = useUserInfo()
 
   if (isLoading) {
     return <div />
   }
 
-  if (!session || isError) {
+  if (!userInfo || isError) {
     return <Navigate to="/login" state={{ from: history.location }} />
   }
